plotter/code/requirements.txt
kafka-python
requests
dash==1.8.0